# Move to using bit sets for ScriptExtension