--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -4,18 +4,24 @@ on: [push]
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
-
+    strategy:
+      matrix:
+        rust:
+          - beta
+          - nightly
     steps:
     - uses: actions/checkout@v1
     - uses: actions-rs/toolchain@v1
       with:
         profile: minimal
-        toolchain: beta
+        toolchain: ${{ matrix.rust }}
         override: true
         components: rustfmt
     - name: Build
       run: cargo build --verbose
     - name: Run tests
       run: cargo test
+    - name: Run benchmarks
+      run: cargo bench --features bench
+      if: startsWith(matrix.rust, 'nightly')