--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "unicode-script"
-version = "0.4.0"
+version = "0.5.0"
 authors = ["Manish Goregaokar <manishsmail@gmail.com>"]
 edition = "2018"
 
@@ -20,9 +20,8 @@ exclude = [ "target/*", "Cargo.lock", "scripts/tmp", "*.txt" ]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-with_std = []
-default_features = ["with_std"]
 rustc-dep-of-std = ['std', 'core', 'compiler_builtins']
+bench = []
 
 [dependencies]
 std = { version = "1.0", package = "rustc-std-workspace-std", optional = true }