Should be much faster and cleaner, since the actual implementation is no longer using a ton of autogenerated enum matches and is instead using bit operations in the handwritten part of the crate.

Fixes #2

Probably won't merge till I add tests.