@Manishearth 2020-01-11
Some tests exist, I kinda want to add more including some manual ones (and maybe pull in quickcheck)