@Manishearth 2020-01-11
I also need to check perf of the important operations