@Manishearth 2020-01-11
Benchmark numbers:

```
New:
test tests::bench_ext_intersection    ... bench:           2 ns/iter (+/- 0)
test tests::bench_ext_to_script       ... bench:           5 ns/iter (+/- 0)
test tests::bench_script_to_ext       ... bench:           1 ns/iter (+/- 0)
test tests::bench_script_intersection ... bench:           3 ns/iter (+/- 0)
test tests::bench_to_vec              ... bench:           4 ns/iter (+/- 0)
test tests::bench_string_ext          ... bench:       7,175 ns/iter (+/- 655)

Old:
test tests::bench_ext_intersection    ... bench:           5 ns/iter (+/- 0)
test tests::bench_ext_to_script       ... bench:           0 ns/iter (+/- 0)
test tests::bench_script_to_ext       ... bench:           0 ns/iter (+/- 0)
test tests::bench_script_intersection ... bench:           7 ns/iter (+/- 0)
test tests::bench_to_vec              ... bench:          14 ns/iter (+/- 1)
test tests::bench_string_ext          ... bench:       7,292 ns/iter (+/- 631)
```

It's better for most operations except for conversions (which is to be expected). The common one -- converting a `Script` to a `ScriptExtension` -- is still pretty fast. The actual lookup operation seems to dominate for matching.